breakdown object into files
